% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_NGT_filter.R
\name{active_NGT_filter}
\alias{active_NGT_filter}
\title{Active learning approach to iteratively filter NGT matrix for missing genotype information.}
\usage{
active_NGT_filter(
  NGT_filter,
  variants_of_interest,
  required_variants,
  required_cells,
  variant_score_cutoff = 0.9,
  cell_score_cutoff = 0.9,
  greedy_scalar = 0.01
)
}
\arguments{
\item{NGT_filter}{required, filtered NGT matrix containing 0,1,2 and NA for cells on rows, and variants in columnns}

\item{variants_of_interest}{required, universe of potential variants that have been prefiltered}

\item{required_variants}{optional, vector of variants that cannot be removed}

\item{required_cells}{optional, vector of cells that cannot be removed}

\item{variant_score_cutoff}{required, convergence value for fraction of cells that a given variant is successfully genotyped for scale is 0-1}

\item{cell_score_cutoff}{required,  convergence value for fraction of variants that a given cell is successfully genotyped for scale is 0-1}

\item{greedy_scalar}{optional, scalar value to speed up computation and remove more cells/variants at each iteration, keep at 0.01 or less}
}
\value{
NGT_subset, a matrix with reduced dimensions from NGT_filter that possesses well represented variants and cells
}
\description{
Active learning approach to iteratively filter NGT matrix for missing genotype information.
}
