% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonograph.R
\name{clonograph}
\alias{clonograph}
\title{Plotting clonographs}
\usage{
clonograph(sce, complete_only = FALSE, color_pal = "Reds", QC_stats = FALSE)
}
\arguments{
\item{sce}{the singleCellExperiment object with the clones, NGT, and architecture.}

\item{complete_only}{logical to decide whether to plot only complete cells or also low quality cells, default=FALSE}

\item{color_pal}{brewer.pal pallete selection.}

\item{QC_stats}{A true/false call to see other diagnostic plots under the clonograph (default=FALSE)}
}
\value{
A clonograph figure
}
\description{
This function creates a plot we call a clonograph. This includes an abundance barplot over a the genotype heatmap. This figure is inspired by upset plots.
}
\examples{
\dontrun{
clonograph(sce,complete_only=TRUE,color_pals="Blues",QC_stats=T)
}
}
