% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_clones.R
\name{enumerate_clones}
\alias{enumerate_clones}
\title{Enumerate clones}
\usage{
enumerate_clones(sce, replicates = 100)
}
\arguments{
\item{sce}{SingleCellExperiment object containing NGT matrix for clone identification.}

\item{replicates}{number of bootstrapping replicates}
}
\value{
updated sce containing a table of clones, modified NGT matrix, and clonal architecture.
}
\description{
This function identifies the 95% confidence interval and a mean value for the abundance of clones in each distribution. The confidence interval uses a bootstrapping approach to find the confidence intervals. Cells are broken down into 2 groups, "Complete" meaning it passed our additional QC and "Other" meaning it didn't pass. Each clone may have both groups.
}
\examples{
\dontrun{
sce<-enumerate_clones(sce, replicates = 500)
}
}
