% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_droplet_size.R
\name{extract_droplet_size}
\alias{extract_droplet_size}
\title{Extract Droplet Size}
\usage{
extract_droplet_size(sce)
}
\arguments{
\item{sce}{singleCellExperiment object}
}
\value{
metadata dataframe which contains droplet_name, dna_size, amplicon, protein_id, and protein_size as well as the NGT matrix
}
\description{
Extract protein library size, dna library size, and amplicon size for all droplets.
}
\examples{
\dontrun{
droplet_metadata<- extract_droplet_size(sce)
}
}
