% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_export.R
\name{fcs_export}
\alias{fcs_export}
\title{FCS Export}
\usage{
fcs_export(sce, slot = "DSB_norm", save_path = "~/sample_test.fcs")
}
\arguments{
\item{sce}{input sce object}

\item{slot}{which protein dataset to use. Options are DSB_norm (default), CLR_norm or raw}

\item{save_path}{path to where fcs file should be saved.}
}
\value{
it will save a file at the specified path.
}
\description{
This function allows for protein data to be exported as an FCS file to be viewed in flow cytometry software.
}
\examples{
\dontrun{
fcs_export(sce,slot="DSB_norm",save_path ="~/sample_fcs_file.fcs")
}
}
