% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lowQC_matrix.R
\name{generate_lowQC_matrix}
\alias{generate_lowQC_matrix}
\title{Extract genotypes for missing data in filtered NGT matrices}
\usage{
generate_lowQC_matrix(NGT_to_fill, complete_NGT, file)
}
\arguments{
\item{NGT_to_fill}{Matrix with missing values that did not pass QC}

\item{file}{path to H5 file containing complete NGT matrix.}

\item{perfect_NGT}{perfect matrix for indexing groups oof cells}
}
\value{
imperfect NGT matrix with cells and variants of interest
}
\description{
Extract genotypes for missing data in filtered NGT matrices
}
