% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_txdb.R
\name{generate_txdb}
\alias{generate_txdb}
\title{generating a transcription data base (TxDb)}
\usage{
generate_txdb(gtf_gz_file_path, folder_saved_directory = NULL)
}
\arguments{
\item{gtf_gz_file_path}{a gtf.gz file for creating a custom txDb and suggested cbioPortalTranscript dataframe.}

\item{folder_saved_directory}{where to save the custom TxDB file and suggested cbioportal_files.}
}
\value{
It returns a list of custom TxDb file and a recommended cbioPortal.
}
\description{
This helper function allows for custom TxDbs to be generated. Note: our annotation method relies on cBioPortal's transcripts as a recommended transcript of interest.
}
\examples{
\dontrun{
custom_txdb <- generate_txdb(panel_bed_path="~/my_bed_file.bed",folder_saved_directory= "~/my_custom_txdb")
}
}
