% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_to_sce.R
\name{loom_to_sce}
\alias{loom_to_sce}
\title{Title}
\usage{
loom_to_sce(
  file,
  GT_cutoff = 35,
  VAF_cutoff = 5,
  DP_cutoff = 10,
  GQ_cutoff = 30,
  AF_cutoff = 25,
  variant_set = NULL
)
}
\arguments{
\item{file}{path to h5 file}

\item{GT_cutoff}{Fraction of cells that are successfully genotyped for initial filtering (default 0.2, meaning 20\%)}

\item{VAF_cutoff}{Fraction of cells that are mutated for initial filtering of variants (default 0.005, meaning 0.05\%)}

\item{DP_cutoff}{minimum number of reads necessary for a reliable genotype call in a single cell (default: 10)}

\item{GQ_cutoff}{minimum genotype quality necessary for a reliable genotype call in a single cell (default: 30)}

\item{AF_cutoff}{Deviation from 0, 50, or 100\% for a reliable call of WT, Het or Hom respectively (default: 25)}
}
\value{
list
}
\description{
Title
}
