% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_filter_NGT.R
\name{quality_filter_NGT}
\alias{quality_filter_NGT}
\title{Filter NGT matrix based on read depth, allele frequency, and genotype quality}
\usage{
quality_filter_NGT(file, NGT, DP_cut = 10, AF_cut = 20, GQ_cut = 20)
}
\arguments{
\item{file}{path to H5 file}

\item{NGT}{NGT object from 'read_tapestri_h5_NGT' function}

\item{DP_cut}{read depth cutoff}

\item{AF_cut}{allele frequency cutoff}

\item{GQ_cut}{genotype quality cutoff#'}
}
\value{
filtered NGT matrix containing NAs for cells that do not pass QC
}
\description{
Filter NGT matrix based on read depth, allele frequency, and genotype quality
}
