% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_output.R
\name{quality_output}
\alias{quality_output}
\title{Produce long form quality metrics for each cell-variant pair for read depth, allele frequency, and genotype quality}
\usage{
quality_output(
  file,
  filter = TRUE,
  input_variants,
  input_cells,
  NGT,
  DP_cut = 10,
  AF_cut = 20,
  GQ_cut = 20
)
}
\arguments{
\item{file}{path to H5 file}

\item{filter}{default=TRUE, determines whether function should output filtered or unfiltered long form matrix containing per cell-variant pair of NGT, DP, AF, GQ metrics}

\item{input_variants}{variants of interest to collect QC data}

\item{input_cells}{cells of interest to subset on}

\item{NGT}{NGT matrix}

\item{DP_cut}{read depth cutoff}

\item{AF_cut}{allele frequency cutoff}

\item{GQ_cut}{genotype quality cutoff}
}
\value{
filtered NGT matrix containing NAs for cells that do not pass QC
}
\description{
Produce long form quality metrics for each cell-variant pair for read depth, allele frequency, and genotype quality
}
