% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tapestri_h5.R
\name{read_tapestri_h5}
\alias{read_tapestri_h5}
\title{Import Tapestri H5 data}
\usage{
read_tapestri_h5(file, VAF_cutoff = 0.005, GT_cutoff = 20)
}
\arguments{
\item{file}{path to h5 file}

\item{VAF_cutoff}{\% of cells that are mutated for initial filtering of variants}

\item{GT_cutoff}{\% of cells that are successfully genotyped for initial filtering}
}
\value{
list
}
\description{
Import Tapestri H5 data
}
