% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tapestri_h5_protein.R
\name{read_tapestri_h5_protein}
\alias{read_tapestri_h5_protein}
\title{Import protein data}
\usage{
read_tapestri_h5_protein(file = file)
}
\arguments{
\item{file}{path to h5 file}
}
\description{
Import protein data
}
