% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_mutations.R
\name{tabulate_mutations}
\alias{tabulate_mutations}
\title{Title}
\usage{
tabulate_mutations(file, GT_cutoff = 0, VAF_cutoff = 0)
}
\arguments{
\item{file}{}

\item{GT_cutoff}{}

\item{VAF_cutoff}{}
}
\description{
Title
}
