% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeMDPs.R
\name{trajectory_of_interest_BSCITE_format}
\alias{trajectory_of_interest_BSCITE_format}
\title{BSCITE Style single trajectory Plot}
\usage{
trajectory_of_interest_BSCITE_format(
  sce,
  trajectory = sce@metadata$Trajectories[[1]],
  save_filename = NULL
)
}
\arguments{
\item{sce}{SingleCellExperiment object containing NGT matrix for clone identification.}

\item{trajectory}{Put in entire Trajectory you are interested in such as 'sce(at)metadata$Trajectories[[1]]'}

\item{save_filename}{file location to save file. Currently, must end it with .html extension.}
}
\value{
interactive graph for the Reinforcement Learning Trajectory
}
\description{
This function produces the same network styled like BSCITE mutation acquisition.
}
\examples{
\dontrun{
trajectory_of_interest_BSCITE_format(sce = sce_subset,trajectory = sce_subset(at)metadata$Trajectories[[2]],save_filename = NULL)
}
}
