% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sce.R
\name{validate_sce}
\alias{validate_sce}
\title{Validate SCE}
\usage{
validate_sce(sce)
}
\arguments{
\item{sce}{The current singleCellExperiment object.}
}
\value{
it outputs a status report directly to the console about processes that have been completed or
}
\description{
This function returns a report of what analyses have been conducted and still are available to be completed.
}
\examples{
\dontrun{

}
}
