% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeMDPs.R
\name{visualize_full_network}
\alias{visualize_full_network}
\title{Visualize Full Network}
\usage{
visualize_full_network(sce, save_filename = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object containing NGT matrix for clone identification.}

\item{save_filename}{file location to save file. Currently, must end it with .html extension.}
}
\value{
interactive graph for the Reinforcement Learning Trajectory
}
\description{
The full network of the entire Markov Decision Process. It will show red circles indicating observed clones, grey circles which are unobserved. The size of the circle represents the abundance of the clones. Black lines indicate a mutation is acquired (WT->HET->HOM), green is ADO(HET->WT or HOM->HET), pink and is forward ADO (HET->HOM, or WT->HET).
}
\examples{
\dontrun{
visualize_full_network(sce = sce_subset,save_filename = NULL)
}
}
