% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variants of interest}
\usage{
annotate_variants(file, annotation_key, txdb, banned, NGT)
}
\arguments{
\item{file}{path to h5 file}

\item{annotation_key}{external data frame containing annotation information}

\item{txdb}{txdb object for genes of interest}

\item{banned}{data.frame of genomic coordinates to be filtered out}

\item{NGT}{filtered NGT matrix from quality_filter_NGT}
}
\value{
variant annotation matrix
}
\description{
Annotate variants of interest
}
