% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variants of interest}
\usage{
annotate_variants(file, panel = NULL, select_variants = NULL)
}
\arguments{
\item{file}{path to h5 file to pull out relevant DNA information}

\item{select_variants}{variants of interest, default is all variants}
}
\value{
variant annotation matrix
}
\description{
Annotate variants of interest
}
