% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_weights.R
\name{attach_weights}
\alias{attach_weights}
\title{Assigns observed counts as weights to the Markov Decision Process}
\usage{
attach_weights(sce, adj_linklist)
}
\arguments{
\item{sce}{The single cell experiment object that carries all data}

\item{adj_linklist}{this is an adjacency list for the theoretical states describing the MDP}
}
\description{
Assigns observed counts as weights to the Markov Decision Process
}
