% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMDP.R
\name{BuildMDP}
\alias{BuildMDP}
\title{The main function builds an adjacency list of the theoretical order of mutations. The same MDP mask can be reused different mutations}
\usage{
BuildMDP(num_mutations, use_ADO = FALSE)
}
\arguments{
\item{num_mutations}{The number of variants we are using (automatically obtained from the Architecture)}
}
\description{
This contains 1 main function and 3 helper functions.
The 3 helper functions aim to:
\enumerate{
\item turn our states to Ternary variables (can be 0(WT), 1(Het), or 2(Hom))
\item a function that finds the difference in states that are < 1 mutation away
\item a function that does bit logic subtraction for the ternary variables
}
}
