% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_confidence_labeling.R
\name{cell_confidence_labeling}
\alias{cell_confidence_labeling}
\title{cell_confidence_labeling
This function provides additional labeling to he quality of the cells based
on dna and protein data.}
\usage{
cell_confidence_labeling(sce)
}
\arguments{
\item{sce}{object}
}
\value{
sce object with updated labeling in the colData to be used later
}
\description{
cell_confidence_labeling
This function provides additional labeling to he quality of the cells based
on dna and protein data.
}
