% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonograph.R
\name{clonograph}
\alias{clonograph}
\title{Plotting clonographs}
\usage{
clonograph(final_sample_summary, complete_only = FALSE, color_pal = "Reds")
}
\arguments{
\item{complete_only}{logical to decide whether to plot only complete cells or also low quality cells, default=FALSE}

\item{color_pal}{brewer.pal pallete selection.}
}
\description{
Plotting clonographs
}
