% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clone_statistics.R
\name{compute_clone_statistics}
\alias{compute_clone_statistics}
\title{Title}
\usage{
compute_clone_statistics(sce, clone_size_cutoff = 10)
}
\arguments{
\item{sce}{input sce object}

\item{clone_size_cutoff}{lower 95\% confidence interval cutoff for clone inclusion default=10}
}
\description{
Title
}
