% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clone_statistics.R
\name{compute_clone_statistics}
\alias{compute_clone_statistics}
\title{Title}
\usage{
compute_clone_statistics(sce, clone_size_cutoff = 10)
}
\arguments{
\item{sce}{}

\item{clone_size_cutoff}{}
}
\description{
Title
}
