% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplex_samples.R
\name{demultiplex_samples}
\alias{demultiplex_samples}
\title{Demultiplex wrapper function to handle splitting samples}
\usage{
demultiplex_samples(
  sce,
  sensitivity_threshold = c(0.01, 1e-04),
  expected_samples = 5
)
}
\arguments{
\item{sce}{a SingleCellExperiment object containing the Clones,Architecture, and NGT matrix}

\item{sensitivity_threshold}{vector of sensitivity thresholds for rows (variants) and columns (cells)}

\item{expected_samples}{the number of clusters or samples you expect}
}
\value{
a dataframe of the cells assigned to a specific cluster.
}
\description{
Demultiplex wrapper function to handle splitting samples
}
