% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_clones.R
\name{enumerate_clones}
\alias{enumerate_clones}
\title{Enumerate clones}
\usage{
enumerate_clones(NGT, variant_metadata, replicates = 100)
}
\arguments{
\item{NGT}{NGT matrix for clone identification. Should lack NA's}

\item{variant_metadata}{metadata}

\item{replicates}{number of bootstrapping replicates}
}
\value{
list containing a table of clones, modified NGT matrix, and clonal architecture.
}
\description{
Enumerate clones
}
