% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_clones.R
\name{enumerate_clones}
\alias{enumerate_clones}
\title{Enumerate clones}
\usage{
enumerate_clones(sce, replicates = 100)
}
\arguments{
\item{sce}{SingleCellExperiment object containing NGT matrix for clone identification.}

\item{replicates}{number of bootstrapping replicates}
}
\value{
updated sce containing a table of clones, modified NGT matrix, and clonal architecture.
}
\description{
Enumerate clones
}
