% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_droplet_size.R
\name{extract_droplet_size}
\alias{extract_droplet_size}
\title{Extract protein library size, dna library size, and amplicon size for all droplets.}
\usage{
extract_droplet_size(sce)
}
\arguments{
\item{sce}{singleCellExperiment Object}

\item{file}{path to h5 file}
}
\value{
metadata matrix
}
\description{
Extract protein library size, dna library size, and amplicon size for all droplets.
}
