% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_export.R
\name{fcs_export}
\alias{fcs_export}
\title{Title}
\usage{
fcs_export(sce, slot = "DSB_norm", save_path = "~/sample_test.fcs")
}
\arguments{
\item{sce}{input sce object}

\item{slot}{which protein dataset to use. Options are DSB_norm (default), CLR_norm or raw}

\item{save_path}{path to where fcs file should be saved}
}
\description{
Title
}
