% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_QC_metrics.R
\name{GetMembership}
\alias{GetMembership}
\title{Get membership quality of cells vs empty}
\usage{
GetMembership(droplet_metadata)
}
\arguments{
\item{droplet_metadata}{}
}
\description{
Get membership quality of cells vs empty
}
