% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdp_Q_learning_with_linklist.R
\name{mdp_Q_learning_with_linklist}
\alias{mdp_Q_learning_with_linklist}
\title{This file run Reinforcement Learning (model-free Q-learning) to evaluate the most likely mutation paths from the data}
\usage{
mdp_Q_learning_with_linklist(df, discount, N)
}
\arguments{
\item{df}{this is the adjacency link link to}

\item{discount}{the factor determines how much future mutations impact current actions. 0 only looks at next step, close to 1 will have a very long horizon time}

\item{N}{Number of iterations, also called episodes to explore the MDP}
}
\value{
A dataframe that contains the MDP adjacency list, normalized Q-values and mutations taken
}
\description{
This file run Reinforcement Learning (model-free Q-learning) to evaluate the most likely mutation paths from the data
}
