% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_protein_data.R
\name{normalize_protein_data}
\alias{normalize_protein_data}
\title{Normalize Protein Data}
\usage{
normalize_protein_data(
  sce,
  metadata,
  method = "dsb",
  detect_IgG = TRUE,
  background_droplets
)
}
\arguments{
\item{NGT}{}
}
\description{
Normalize Protein Data
}
