% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_AF_mat.R
\name{optimize_matrix}
\alias{optimize_matrix}
\title{Title}
\usage{
optimize_matrix(sce, sensitivity_threshold = c(0.01, 0.001))
}
\arguments{
\item{sce}{sce file}

\item{sensitivity_threshold}{vector of sensitivity thresholds for rows (variants) and columns (cells)}
}
\description{
Title
}
