% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDNA_CN_H5.R
\name{readDNA_CN_H5}
\alias{readDNA_CN_H5}
\title{This function generates the Copy Number by determining the ploidy of each mutation}
\usage{
readDNA_CN_H5(sce, reference_cells = NULL)
}
\arguments{
\item{sce}{the single cell experiment object}

\item{reference_cells}{set cells that will be used to determine allele dropout and false postive rates}
}
\description{
This function generates the Copy Number by determining the ploidy of each mutation
}
