% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestri_h5_to_sce.R
\name{tapestri_h5_to_sce}
\alias{tapestri_h5_to_sce}
\title{Import Tapestri H5 data and extract genotype matrix}
\usage{
tapestri_h5_to_sce(
  file,
  GT_cutoff = 35,
  VAF_cutoff = 5,
  DP_cutoff = 10,
  GQ_cutoff = 30,
  AF_cutoff = 25,
  variant_set = NULL,
  final_mutation_set = NULL,
  protein = TRUE,
  return_variants_only = FALSE
)
}
\arguments{
\item{file}{path to the h5 file}

\item{GT_cutoff}{Fraction of cells that are successfully genotyped for initial filtering (default 0.2, meaning 20\%)}

\item{VAF_cutoff}{Fraction of cells that are mutated for initial filtering of variants (default 0.005, meaning 0.05\%)}

\item{DP_cutoff}{minimum number of reads necessary for a reliable genotype call in a single cell (default: 10)}

\item{GQ_cutoff}{minimum genotype quality necessary for a reliable genotype call in a single cell (default: 30)}

\item{AF_cutoff}{Deviation from 0, 50, or 100\% for a reliable call of WT, Het or Hom respectively (default: 25)}

\item{variant_set}{character vector of variants to be included in the format output by mission bio.}

\item{protein}{logical, whether protein data should be included. default=TRUE}

\item{return_variants_only}{logical,}
}
\value{
a single cell experiment object containing the genotyping matrix, allele frequency table, annotation table,
}
\description{
Import Tapestri H5 data and extract genotype matrix
}
