% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_analysis.R
\name{trajectory_analysis}
\alias{trajectory_analysis}
\title{Run Trajectory Analysis after extraction from SingleCellExperiment object}
\usage{
trajectory_analysis(sce, use_ADO = TRUE)
}
\arguments{
\item{sce}{is the SingleCellExperiment Object}
}
\value{
RL_analysis this returns a paths with metadata about the mutation order
}
\description{
Run Trajectory Analysis after extraction from SingleCellExperiment object
}
